+++ cli/src/index.ts
@@ -23,7 +23,7 @@ import { AstarAPI } from "./lib/api";
 import { c } from "./lib/ui";
 import { resolve } from "path";
 
-export const VERSION = "0.0.70";
+export const VERSION = "0.0.71";
 
 async function showDashboard() {
   const status = await getAuthStatus();